Extend Message class with helper methods.
